test: adding buffer manager tests

+++ inc/SpctFxParameters.h
@@ -1,3 +1,10 @@
+/**
+ * Author: Lucas Scheidt
+ * Date: 28.12.24
+ *
+ * Just a simple struct representing all possible FX parameters that are tweakable from external. The plugin is intended
+ * to be updated with a set of such parameters
+ */
 #pragma once
 #include "SpctDomainSpecific.h"
 
@@ -6,9 +13,9 @@ namespace LBTS::Spectral
 struct FxParameters
 {
     OscWaveform waveform_selection;
-    float filter_cutoff;
-    float fft_threshold;
-    float frequency_offset;
+    double filter_cutoff;
+    double fft_threshold;
+    double frequency_offset;
     float gain;
     uint16_t glide_steps;
     size_t voices;
